Cleaned the boiler plate then threw it away. ;-)
