+++ README.rst
@@ -1,3 +1,4 @@
 Data Catalogue loading utilities
+================================
 
 Expects a config file to be located at ~/.dc
