+++ setup.py
@@ -1,32 +1,19 @@
 import re
 import os
-
 from setuptools import setup
+from dc import __version__
 
 
-here = os.path.abspath(os.path.dirname(__file__))
-README = open(os.path.join(here, 'README.md')).read()
+README = open('README.md').read()
 
-
-VERSION_FILE = os.path.join(here, "dc/_version.py")
-verstrline = open(VERSION_FILE, "rt").read()
-VSRE = r'^__version__ = [\'"]([^\'"]*)[\'"]'
-mo = re.search(VSRE,  verstrline, re.M)
-if mo:
-    VERSION = mo.group(1)
-else:
-    raise RuntimeError("Unable to find version string in {0}".format(VERSION_FILE))
-
-os.chdir(here)
 
 setup(
     name = "dc",
-    version = VERSION,
+    version = __version__,
     url = "https://github.com/davidmiller/dc",
     description = "Data Catalogue",
     long_description = README,
-    # install_requires = [
-    #     ],
+    install_requires = ['ffs', ],
     classifiers = [
         "Programming Language :: Python",
         "Programming Language :: Python :: 2.7",
